Dallas temperature - working on setup

+++ sensors_avr.h
@@ -21,7 +21,7 @@
 
 int sensors_vin();
 long sensors_pressure();
-long sensors_temperature();
+int sensors_temperature();
 void sensor_setup();
 
 #endif // ifndef __SENSORS_AVR_H__
